Improved schubert algorithm but fixing bugs
I improved the Schubert algorithm to
1. only do one expansion at a time so it can use all the available info
2. Expand the point with the minimum center value (rather than minimum lower bound)

Unfortunately, this messed up the error calculation. I thought I fixed
it, but the result takes an awfully long time to run, so I split out
the main loop so I can trace it - here's where I'd want a
debugger. But it is late, so I'll have to continue this at the next
opportunity.